requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
google-genai==0.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
